+++ scripts/carride/gui.gd
@@ -11,22 +11,24 @@ var diff_time: int
 var swipe_start: Vector2
 var stop_processing := false
 
-@onready var speedometer: Label = $HBoxContainer3/LabelSpeed
+@onready var speedometer: Label = $SpeedContainer/LabelSpeed
 @onready var game_over_panel := $PanelContainer
 @onready var score_label := $PanelContainer/MarginContainer/VBoxContainer/ScoreLabel
 @onready var play_again_button := $PanelContainer/MarginContainer/VBoxContainer/PlayAgainBtn
 
 
-func _ready():
+func _ready() -> void:
 	if SettingsBus.easier_font:
 		label_time.label_settings.font_size = 32
 	diff_time = 0
 	SignalBus.enable_touchscreen_vbuttons.connect(_enable_vbuttons)
+	SignalBus.cr_speedometer_value.connect(_cr_speedometer_value)
+	_cr_speedometer_value(SettingsBus.cr_speedometer_label)
 	if SettingsBus.touchscreen_control == SettingsBus.TouchscreenControlMode.VBUTTONS:
 		_enable_vbuttons(true)
 
 
-func _process(_delta):
+func _process(_delta: float) -> void:
 	if not stop_processing:
 		diff_time = owner.time
 		var miliseconds: int = diff_time % 1000
@@ -36,18 +38,18 @@ func _process(_delta):
 		speedometer.text = ("%.3f" % [owner.speed])
 
 
-func update_points():
+func update_points() -> void:
 	label_milk.text = str(owner.milks)
 
 
-func game_over():
+func game_over() -> void:
 	stop_processing = true
 	score_label.text = str(owner.final_score)
 	play_again_button.grab_focus()
 	game_over_panel.show()
 
 
-func _gui_input(event):
+func _gui_input(event: InputEvent) -> void:
 	if event is InputEventScreenTouch:
 		if SettingsBus.touchscreen_control == SettingsBus.TouchscreenControlMode.TAP:
 			if event.pressed and event.index == 0:
@@ -62,7 +64,7 @@ func _gui_input(event):
 				_calculate_swipe(event.get_position())
 
 
-func _calculate_swipe(swipe_end: Vector2):
+func _calculate_swipe(swipe_end: Vector2) -> void:
 	if swipe_start == null:
 		return
 	var swipe = swipe_end - swipe_start
@@ -73,24 +75,31 @@ func _calculate_swipe(swipe_end: Vector2):
 			touchscreen_move.emit(true)
 
 
-func _enable_vbuttons(yes: bool):
+func _enable_vbuttons(yes: bool) -> void:
 	if yes:
 		$VButtonsContainer.show()
 	else:
 		$VButtonsContainer.hide()
 
 
-func _on_play_again_btn_pressed():
+func _cr_speedometer_value(yes: bool) -> void:
+	if yes:
+		$SpeedContainer.show()
+	else:
+		$SpeedContainer.hide()
+
+
+func _on_play_again_btn_pressed() -> void:
 	get_tree().change_scene_to_file("res://scenes/carride.tscn")
 
 
-func _on_menu_btn_pressed():
+func _on_menu_btn_pressed() -> void:
 	get_tree().change_scene_to_file("res://scenes/mainMenu.tscn")
 
 
-func _on_up_button_down():
+func _on_up_button_down() -> void:
 	touchscreen_move.emit(true)
 
 
-func _on_down_button_down():
+func _on_down_button_down() -> void:
 	touchscreen_move.emit(false)